v0.0.3 Рефакторинг и второй прототип
