+++ Cargo.toml
@@ -10,3 +10,5 @@ rust_decimal = "1.34.3"
 chrono = "0.4.35"
 rust_decimal_macros = "1.34.2"
 uuid = { version = "1.8.0", features = ["v4"] }
+serde = "1.0.210"
+serde_yaml = "0.9.34"
