+++ src/main.rs
@@ -1,59 +1,73 @@
 use chrono::Utc;
 use rust_decimal_macros::dec;
+use uuid::Uuid;
 
 use anna_ivanovna;
-use anna_ivanovna::distributor::{
-    Absolute, Currency, Income, IncomeSource, Money, Percentage, Plan, Target,
-};
+use anna_ivanovna::distribute::{distribute, Income};
+use anna_ivanovna::finance::{Currency, Money, Percentage};
+use anna_ivanovna::planning::planning::{Expense, ExpenseValue, IncomeSource, Plan};
 
 fn main() {
     let zp = IncomeSource::new(
         "Зарплата".to_string(),
         Money::new(dec!(100000), Currency::RUB),
     );
-    let rent = Absolute::new("Аренда".to_string(), Money::new(dec!(25000), Currency::RUB));
-    let food = Absolute::new("Еда".to_string(), Money::new(dec!(17345), Currency::RUB));
-    let health = Absolute::new(
+
+    let rent = Expense::new(
+        "Аренда".to_string(),
+        ExpenseValue::MONEY {
+            value: Money::new_rub(dec!(25000)),
+        },
+    );
+    let food = Expense::new(
+        "Еда".to_string(),
+        ExpenseValue::MONEY {
+            value: Money::new(dec!(17345), Currency::RUB),
+        },
+    );
+    let health = Expense::new(
         "Здоровье".to_string(),
-        Money::new(dec!(10000), Currency::RUB),
+        ExpenseValue::MONEY {
+            value: Money::new(dec!(10000), Currency::RUB),
+        },
     );
-    let mortgage = Absolute::new(
+    let mortgage = Expense::new(
         "Ипотека".to_string(),
-        Money::new(dec!(10000), Currency::RUB),
+        ExpenseValue::MONEY {
+            value: Money::new(dec!(10000), Currency::RUB),
+        },
     );
-    let home_service = Absolute::new(
+    let home_service = Expense::new(
         "Коммуналка".to_string(),
-        Money::new(dec!(6000), Currency::RUB),
+        ExpenseValue::MONEY {
+            value: Money::new(dec!(6000), Currency::RUB),
+        },
     );
-    let bag_month = Absolute::new(
+    let bag_month = Expense::new(
         "Подушка на месяц".to_string(),
-        Money::new(dec!(5000), Currency::RUB),
+        ExpenseValue::MONEY {
+            value: Money::new(dec!(5000), Currency::RUB),
+        },
     );
 
-    let auto = Target::new(
+    let auto = Expense::new(
         "Авто".to_string(),
-        Money::new(dec!(1_000_000), Currency::RUB),
-        Percentage::from_int(10),
+        ExpenseValue::RATE {
+            value: Percentage::from(dec!(26.6599)),
+        },
     );
 
     let plan = Plan::try_build(
+        Uuid::new_v4(),
         vec![zp.clone()],
-        vec![
-            rent,
-            food,
-            health,
-            mortgage,
-            home_service,
-            bag_month,
-        ],
-        vec![auto],
+        vec![rent, food, health, mortgage, home_service, bag_month, auto],
     )
         .unwrap();
 
     let income = Income::new(
         zp.clone(),
-        Utc::now(),
-        Money::new(dec!(50000), Currency::RUB),
+        Money::new_rub(dec!(50000)),
+        Utc::now().date_naive(),
     );
-    print!("{}", plan.new_income(&income).unwrap());
+    print!("{}", distribute(plan, income).unwrap());
 }
